requirements.txt
tqdm==4.59.0
numpy==1.20.1
tensorboard==2.4.1
Pillow==8.1.2
opencv-python==4.5.1.48
torch==1.4.0
torchvision==0.5.0
